refactor(cli): group helper provider commands into provider sub-package
