+++ cmd/internal/helper/provider/check_provider_update.go
@@ -1,4 +1,4 @@
-package helper
+package provider
 
 import (
 	"bytes"
