+++ cmd/internal/helper/provider/get_provider_name.go
@@ -1,4 +1,4 @@
-package helper
+package provider
 
 import (
 	"bytes"
